+++ example/app/app.component.ts
@@ -7,4 +7,5 @@ import {Component} from "@angular/core";
     templateUrl: "./app.component.html"
 })
 
-export class AppComponent {}
+export class AppComponent {
+}