+++ modules/ui-semantic/src/dynamic-form-semantic.component.ts
@@ -1,12 +1,11 @@
-import {Component, forwardRef, Input, ContentChild, TemplateRef} from "@angular/core";
-import {REACTIVE_FORM_DIRECTIVES, FormGroup} from "@angular/forms";
+import {Component, Input, ContentChild, TemplateRef} from "@angular/core";
+import {FormGroup} from "@angular/forms";
 import {DynamicFormControlComponent, DynamicFormControlModel} from "@ng2-dynamic-forms/core";
 
 export const DYNAMIC_FORM_UI_SEMANTIC = "SEMANTIC";
 
 @Component({
 
-    directives: [REACTIVE_FORM_DIRECTIVES],
     moduleId: module.id,
     selector: "dynamic-form-semantic-control",
     templateUrl: "./dynamic-form-semantic.component.html"